+++ NEWS.md
@@ -9,6 +9,9 @@
 * `viz_contributions()` displays the expected average contribution.
 * `pca()` gained a new argument to specify supplementary qualitative variables.
 
+## Breaking changes
+* `plot_*()` methods are now defunct (deprecated in v0.4.0).
+
 # dimensio 0.5.0
 ## New classes and methods
 * Add `cdt()` to compute the complete disjunctive table of a factor table.