+++ NEWS.md
@@ -1,4 +1,4 @@
-# dimensio 0.5.0.9000
+# dimensio 0.6.0
 ## New classes and methods
 * Add `predict()` method for MCA.